wip: Factor out helpers more
